requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
beautifulsoup4
requests
py-flags
